v0.1a
First implementation of word getter!

picks a random word from a list of the 47,000 most common English words. The list is not curated so there's some duds in there. I also have the whole thing loaded in as a static array. maybe I'll eventually update this to use a database or something.